+++ front/src/components/actions/auth.tsx
@@ -1,0 +1,11 @@
+import Logout from "@mui/icons-material/Logout";
+import store from "../../state/store";
+import { unsetAccessToken } from "../../state/userSlice";
+import Action from "./action";
+
+export const LogoutAction: Action = {
+	label: 'Logout',
+	icon: <Logout/>,
+	href: '/',
+	onClick: () => store.dispatch(unsetAccessToken())
+};
